i18n of zxcvbn
