+++ src/locales/ar-SA.json
@@ -1299,5 +1299,57 @@
 		"error": "فشل التحقق",
 		"errorMessage": "فشل التحقق: ${error}",
 		"wasm_disabled": "تمكين WASM للتحقق بشكل أسرع"
+	},
+	"zxcvbn": {
+		"warnings": {
+			"straightRow": "مفاتيح مستمرة على لوحة المفاتيح يتم تخمينها بسهولة.",
+			"keyPattern": "من السهل تخمين أنماط لوحة المفاتيح القصيرة.",
+			"simpleRepeat": "يتم تخمين الشخصيات المتكررة مثل \"AAA\" بسهولة.",
+			"extendedRepeat": "يتم تخمين أنماط الشخصية المتكررة مثل \"ABCABCABC\" بسهولة.",
+			"sequences": "يتم تخمين تسلسلات الأحرف الشائعة مثل \"ABC\" بسهولة.",
+			"recentYears": "من السهل تخمين السنوات الأخيرة.",
+			"dates": "التاريخ يمكن تخمينه بسهولة.",
+			"topTen": "كلمة المرور هذه مُستخدمة بكثرة.",
+			"topHundred": "هذه كلمة مرور تستخدم غالبًا.",
+			"common": "هذه كلمة مرور شائعة الاستخدام.",
+			"similarToCommon": "هذا يشبه كلمة المرور الشائعة الاستخدام.",
+			"wordByItself": "كلمة واحدة يمكن تخمينها بسهولة.",
+			"namesByThemselves": "اسم واحد أو الأخير أو العائلة يتم تخمينه بسهولة.",
+			"commonNames": "من السهل تخمين الأسماء الأولى الشائعة والأسماء الأخيرة.",
+			"userInputs": "يجب ألا تحتوي كلمة المرور على أي بيانات شخصية أو ذات صلة بالصفحة.",
+			"pwned": "لقد تعرضت كلمة المرور الخاصة بك في خرق بيانات الإنترنت."
+		},
+		"suggestions": {
+			"l33t": "تجنب استخدام بدائل الحروف التي يمكن التنبؤ بها ، مثل استخدام \"@\" بدلاً من \"A\".",
+			"reverseWords": "تجنب استخدام الإملاء العكسي للكلمات الشائعة.",
+			"allUppercase": "استخدم الأحرف الكبيرة لبعض الحروف وليس كلها.",
+			"capitalization": "استخدم الأحرف الكبيرة لأكثر من حرف واحد.",
+			"dates": "تجنب استخدام التواريخ والسنوات المتعلقة بك.",
+			"recentYears": "تجنب استخدام آخر السنوات.",
+			"associatedYears": "تجنب استخدام السنوات المتعلقة بك.",
+			"sequences": "تجنب تسلسل الأحرف الشائعة.",
+			"repeated": "تجنب الكلمات والشخصيات المتكررة.",
+			"longerKeyboardPattern": "استخدم وضع لوحة المفاتيح الأطول وتغيير اتجاهات الإدخال عدة مرات.",
+			"anotherWord": "أضف المزيد من الكلمات غير المألوفة.",
+			"useWords": "استخدم كلمات متعددة ، ولكن تجنب العبارات الشائعة.",
+			"noNeed": "لا تحتاج إلى استخدام الرموز أو الأرقام أو الحروف الرأسمالية لإنشاء كلمات مرور قوية.",
+			"pwned": "إذا كنت تستخدم كلمة المرور هذه في مكان آخر، فننصحك بتغييرها."
+		},
+		"timeEstimation": {
+			"ltSecond": "أقل من ثانية",
+			"second": "{base} ثانية",
+			"seconds": "{base} ثوانٍ",
+			"minute": "{base} دقيقة",
+			"minutes": "{base} دقائق",
+			"hour": "{base} ساعة",
+			"hours": "{base} ساعات",
+			"day": "{base} يوم",
+			"days": "{base} أيام",
+			"month": "{base} شهر",
+			"months": "{base} أشهر",
+			"year": "{base} سنة",
+			"years": "{base} سنوات",
+			"centuries": "قرون"
+		}
 	}
 }
