+++ src/locales/hi-IN.json
@@ -1299,5 +1299,57 @@
 		"error": "सत्यापन विफल",
 		"errorMessage": "सत्यापन विफल: ${error}",
 		"wasm_disabled": "तेजी से सत्यापन के लिए WASM सक्षम करें"
+	},
+	"zxcvbn": {
+		"warnings": {
+			"straightRow": "कीबोर्ड पर निरंतर कुंजियों को आसानी से अनुमान लगाया जाता है।",
+			"keyPattern": "एक छोटे कीबोर्ड पैटर्न का अनुमान लगाना आसान है।",
+			"simpleRepeat": "\"एएए\" जैसे दोहराए गए पात्रों को आसानी से अनुमान लगाया जाता है।",
+			"extendedRepeat": "\"ABCABCABC\" जैसे बार -बार चरित्र पैटर्न आसानी से अनुमानित हैं।",
+			"sequences": "\"एबीसी\" जैसे सामान्य चरित्र अनुक्रमों का आसानी से अनुमान लगाया जाता है।",
+			"recentYears": "हाल के वर्षों का अनुमान लगाना आसान है।",
+			"dates": "दिनांक आसानी से अनुमान लगाया जाता है।",
+			"topTen": "यह एक अति उपयोग किया गया पासवर्ड है।",
+			"topHundred": "यह एक पासवर्ड है जिसका उपयोग अक्सर किया जाता है।",
+			"common": "यह आमतौर पर इस्तेमाल किया जाने वाला पासवर्ड है।",
+			"similarToCommon": "यह आमतौर पर उपयोग किए जाने वाले पासवर्ड के समान है।",
+			"wordByItself": "एक शब्द को आसानी से अनुमान लगाया जाता है।",
+			"namesByThemselves": "एक एकल पहला या अंतिम नाम आसानी से अनुमान लगाया जाता है।",
+			"commonNames": "आम पहले नाम और अंतिम नामों का अनुमान लगाना आसान है।",
+			"userInputs": "पासवर्ड में कोई व्यक्तिगत या पृष्ठ-संबंधित डेटा नहीं होना चाहिए।",
+			"pwned": "आपका पासवर्ड एक इंटरनेट डेटा ब्रीच में उजागर किया गया है।"
+		},
+		"suggestions": {
+			"l33t": "पूर्वानुमानित पत्र प्रतिस्थापन का उपयोग करने से बचें, जैसे कि \"ए\" के बजाय \"@\" का उपयोग करना।",
+			"reverseWords": "सामान्य शब्दों की रिवर्स स्पेलिंग का उपयोग करने से बचें।",
+			"allUppercase": "सभी पत्रों के बजाय कुछ को कैपिटल करें।",
+			"capitalization": "केवल पहले अक्षर को ही नहीं, बल्कि अन्य अक्षरों को भी बड़ा लिखें।",
+			"dates": "आपसे संबंधित तारीखों और वर्षों का उपयोग करने से बचें।",
+			"recentYears": "सबसे हाल के वर्षों का उपयोग करने से बचें।",
+			"associatedYears": "आपसे संबंधित वर्षों का उपयोग करने से बचें।",
+			"sequences": "सामान्य चरित्र अनुक्रमों से बचें।",
+			"repeated": "दोहराए जाने वाले शब्दों और पात्रों से बचें।",
+			"longerKeyboardPattern": "लंबे समय तक कीबोर्ड मोड का उपयोग करें और कई बार इनपुट दिशाएं बदलें।",
+			"anotherWord": "अधिक असामान्य शब्द जोड़ें।",
+			"useWords": "कई शब्दों का उपयोग करें, लेकिन सामान्य वाक्यांशों से बचें।",
+			"noNeed": "मजबूत पासवर्ड बनाने के लिए आपको प्रतीकों, संख्याओं या बड़े अक्षरों का उपयोग करने की आवश्यकता नहीं है।",
+			"pwned": "यदि आप इस पासवर्ड का उपयोग कहीं और करते हैं, तो आपको परिवर्तन करने की सलाह दी जाती है।"
+		},
+		"timeEstimation": {
+			"ltSecond": "एक सेकंड से कम",
+			"second": "{base} सेकंड",
+			"seconds": "{base} सेकंड",
+			"minute": "{base} मिनट",
+			"minutes": "{base} मिनट",
+			"hour": "{base} घंटे",
+			"hours": "{base} घंटे",
+			"day": "{base} दिन",
+			"days": "{base} दिन",
+			"month": "{base} महीना",
+			"months": "{base} महीने",
+			"year": "{base} वर्ष",
+			"years": "{base} वर्ष",
+			"centuries": "सदियों"
+		}
 	}
 }
