+++ src/locales/ru-RU.json
@@ -1299,5 +1299,57 @@
 		"error": "Проверка не удалась",
 		"errorMessage": "Проверка не удалась: ${error}",
 		"wasm_disabled": "Включите WASM для более быстрой проверки."
+	},
+	"zxcvbn": {
+		"warnings": {
+			"straightRow": "Непрерывные клавиши на клавиатуре легко угадаются.",
+			"keyPattern": "Короткий шаблон клавиатуры легко угадать.",
+			"simpleRepeat": "Повторяющиеся символы, такие как «AAA», легко угадать.",
+			"extendedRepeat": "Повторные шаблоны символов, такие как «Abcabcabc», легко угадаются.",
+			"sequences": "Общие последовательности символов, такие как «ABC», легко угадаются.",
+			"recentYears": "Последние годы легко догадаться.",
+			"dates": "Дата легко догадается.",
+			"topTen": "Это чрезмерный пароль.",
+			"topHundred": "Это пароль, который часто используется.",
+			"common": "Это обычно используемый пароль.",
+			"similarToCommon": "Это похоже на широко используемый пароль.",
+			"wordByItself": "Одно слово легко угадается.",
+			"namesByThemselves": "Однозначное имя или фамилия легко догадается.",
+			"commonNames": "Общие имена и фамилии легко догадаться.",
+			"userInputs": "Пароль не должен содержать каких-либо личных или страничных данных.",
+			"pwned": "Ваш пароль был скомпрометирован в интернет -нарушении данных."
+		},
+		"suggestions": {
+			"l33t": "Избегайте использования предсказуемых буквенных замен, таких как использование «@» вместо «A».",
+			"reverseWords": "Избегайте использования обратного написания общих слов.",
+			"allUppercase": "Используйте заглавные буквы для некоторых, а не для всех букв.",
+			"capitalization": "Используйте больше одной заглавной буквы.",
+			"dates": "Избегайте использования дат и лет, связанных с вами.",
+			"recentYears": "Избегайте использования последних лет.",
+			"associatedYears": "Избегайте использования лет, связанных с вами.",
+			"sequences": "Избегайте общих последовательностей символов.",
+			"repeated": "Избегайте повторяющихся слов и персонажей.",
+			"longerKeyboardPattern": "Используйте более длинный режим клавиатуры и измените направления ввода несколько раз.",
+			"anotherWord": "Добавьте более необычные слова.",
+			"useWords": "Используйте несколько слов, но избегайте общих фраз.",
+			"noNeed": "Вам не нужно использовать символы, цифры или заглавные буквы для создания прочных паролей.",
+			"pwned": "Если вы используете этот пароль в другом месте, вам рекомендуется внести изменения."
+		},
+		"timeEstimation": {
+			"ltSecond": "Менее чем секунду",
+			"second": "{base} секунд",
+			"seconds": "{base} секунд",
+			"minute": "{base} мин",
+			"minutes": "{base} мин",
+			"hour": "{base} час",
+			"hours": "{base} часы",
+			"day": "{base} день",
+			"days": "{base} день",
+			"month": "{base} месяц",
+			"months": "{base} месяц",
+			"year": "{base} год",
+			"years": "{base} лет",
+			"centuries": "Столетия"
+		}
 	}
 }
