+++ src/locales/uk-UA.json
@@ -1299,5 +1299,57 @@
 		"error": "Перевірка не вдалася",
 		"errorMessage": "Перевірка не вдалася: ${error}",
 		"wasm_disabled": "Увімкніть WASM для швидшої перевірки"
+	},
+	"zxcvbn": {
+		"warnings": {
+			"straightRow": "Безперервні клавіші на клавіатурі легко здогадуються.",
+			"keyPattern": "Короткий візерунок клавіатури легко здогадуватися.",
+			"simpleRepeat": "Неодноразові символи, як \"AAA\", легко здогадуються.",
+			"extendedRepeat": "Неодноразові моделі символів, як \"Abcabcabc\", легко здогадуються.",
+			"sequences": "Поширені послідовності символів на кшталт \"ABC\" легко здогадуються.",
+			"recentYears": "Останні роки легко вгадати.",
+			"dates": "Дату легко здогадуються.",
+			"topTen": "Це надмірно використаний пароль.",
+			"topHundred": "Це пароль, який часто використовується.",
+			"common": "Це часто використовуваний пароль.",
+			"similarToCommon": "Це схоже на загально використовуваний пароль.",
+			"wordByItself": "Одне слово легко здогадується.",
+			"namesByThemselves": "Одне ім’я або прізвище легко здогадується.",
+			"commonNames": "Загальні імена та прізвища легко здогадуються.",
+			"userInputs": "Пароль не повинен містити жодних особистих або сторінок, пов'язаних з сторінкою.",
+			"pwned": "Ваш пароль був викритий у порушенні даних в Інтернеті."
+		},
+		"suggestions": {
+			"l33t": "Уникайте використання передбачуваних заміни літер, наприклад, використання \"@\" замість \"a\".",
+			"reverseWords": "Уникайте використання зворотного написання загальних слів.",
+			"allUppercase": "Використовуйте великі літери для деяких, а не для всіх букв.",
+			"capitalization": "Використовуйте більше однієї великої літери.",
+			"dates": "Уникайте використання дати та років, пов’язаних з вами.",
+			"recentYears": "Уникайте використання останніх років.",
+			"associatedYears": "Уникайте використання років, пов’язаних з вами.",
+			"sequences": "Уникайте загальних послідовностей символів.",
+			"repeated": "Уникайте повторюваних слів та персонажів.",
+			"longerKeyboardPattern": "Використовуйте довший режим клавіатури та змініть вказівки введення кілька разів.",
+			"anotherWord": "Додайте більше нечасті слова.",
+			"useWords": "Використовуйте кілька слів, але уникайте поширених фраз.",
+			"noNeed": "Для створення сильних паролів вам не потрібно використовувати символи, цифри чи капітальні літери.",
+			"pwned": "Якщо ви використовуєте цей пароль в іншому місці, вам рекомендується внести зміни."
+		},
+		"timeEstimation": {
+			"ltSecond": "Менше секунди",
+			"second": "{base} секунди",
+			"seconds": "{base} секунди",
+			"minute": "{base} хв",
+			"minutes": "{base} хв",
+			"hour": "{base} години",
+			"hours": "{base} години",
+			"day": "{base} день",
+			"days": "{base} день",
+			"month": "{base} місяць",
+			"months": "{base} місяць",
+			"year": "{base} рік",
+			"years": "{base} років",
+			"centuries": "Століття"
+		}
 	}
 }
