+++ src/locales/zh-CN.json
@@ -1299,5 +1299,57 @@
 		"error": "验证失败",
 		"errorMessage": "验证失败：${error}",
 		"wasm_disabled": "启用 WASM 可加快验证速度"
+	},
+	"zxcvbn": {
+		"warnings": {
+			"straightRow": "键盘上的连续按键很容易被猜到。",
+			"keyPattern": "简短的键盘序列很容易被猜到。",
+			"simpleRepeat": "像“aaa”这样的重复字符很容易被猜到。",
+			"extendedRepeat": "像“abcabcabc”这样的重复字符模式很容易被猜到。",
+			"sequences": "像“abc”这样的常见字符序列很容易被猜到。",
+			"recentYears": "最近的年份很容易被猜到。",
+			"dates": "日期很容易被猜到。",
+			"topTen": "这是一个被过度使用的密码。",
+			"topHundred": "这是一个经常被使用的密码。",
+			"common": "这是一个常用的密码。",
+			"similarToCommon": "这与一个常用的密码相似。",
+			"wordByItself": "单个单词很容易被猜到。",
+			"namesByThemselves": "单个名字或姓氏很容易被猜到。",
+			"commonNames": "常见的名字和姓氏很容易被猜到。",
+			"userInputs": "密码中不应包含任何个人或与页面相关的数据。",
+			"pwned": "您的密码已在互联网数据泄露事件中被曝光。"
+		},
+		"suggestions": {
+			"l33t": "避免使用可预测的字母替换，例如用“@”代替“a”。",
+			"reverseWords": "避免使用常见单词的倒序拼写。",
+			"allUppercase": "将部分而非全部字母大写。",
+			"capitalization": "不仅仅将首字母大写。",
+			"dates": "避免使用与您相关的日期和年份。",
+			"recentYears": "避免使用最近的年份。",
+			"associatedYears": "避免使用与您相关的年份。",
+			"sequences": "避免使用常见的字符序列。",
+			"repeated": "避免重复的单词和字符。",
+			"longerKeyboardPattern": "使用更长的键盘序列并多次改变输入方向。",
+			"anotherWord": "添加更多不常见的单词。",
+			"useWords": "使用多个单词，但避免使用常见的短语。",
+			"noNeed": "您无需使用符号、数字或大写字母即可创建强密码。",
+			"pwned": "如果您在其他地方使用此密码，建议您进行更改。"
+		},
+		"timeEstimation": {
+			"ltSecond": "不到一秒",
+			"second": "{base}秒",
+			"seconds": "{base}秒",
+			"minute": "{base}分钟",
+			"minutes": "{base}分钟",
+			"hour": "{base}小时",
+			"hours": "{base}小时",
+			"day": "{base}天",
+			"days": "{base}天",
+			"month": "{base}月",
+			"months": "{base}月",
+			"year": "{base}年",
+			"years": "{base}年",
+			"centuries": "几个世纪"
+		}
 	}
 }
